merge_single

+++ cifar10_double.py
@@ -403,7 +403,7 @@ def main():
     parser.add_argument('--data_transform', type=bool, default=False)
     parser.add_argument('--kn_init', type=bool, default=False)
     parser.add_argument('--align_loss', type=str, default=None)
-    parser.add_argument('--weight_align_factor', type=int, default=250, )
+    parser.add_argument('--weight_align_factor', type=float, default=250.0, )
     parser.add_argument('--lr', type=float, default=1e-3, metavar='LR',
                         help='learning rate (default: 1.0)')
     parser.add_argument('--gamma', type=float, default=0.7,
